Fix/gauzy server settings (#6658)
* feat: automatic unsubscribe

* fix: remove wrong provider

* fix: interceptors order

* fix: auth guard event

* fix: transform in awaitable event

* fix: use invokable method

* fix: improve logout from desktop timer

* feat: add missing provider imports

* fix: handle logout event from gauzy ui

* feat: make api call only if it desktop timer

* fix: infinite loading restart

* fix: mark optional auth strategy.

+++ packages/desktop-ui-lib/src/lib/settings/settings.component.ts
@@ -5,11 +5,13 @@ import {
 	ViewChild,
 	ElementRef,
 	NgZone,
-	AfterViewInit
+	AfterViewInit,
+	Optional
 } from '@angular/core';
 import { TimeTrackerService } from '../time-tracker/time-tracker.service';
 import { NbDialogService, NbToastrService } from '@nebular/theme';
 import { ElectronService } from '../electron/services';
+import { BehaviorSubject, Observable, filter, tap, firstValueFrom } from 'rxjs';
 import { BehaviorSubject, Observable, filter, tap, firstValueFrom } from 'rxjs';
 import { AboutComponent } from '../dialogs/about/about.component';
 import { SetupService } from '../setup/setup.service';
@@ -17,7 +19,10 @@ import * as moment from 'moment';
 import { ToastrNotificationService } from '../services';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { AuthStrategy } from '../auth';
+import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
+import { AuthStrategy } from '../auth';
 
+@UntilDestroy({ checkProperties: true })
 @UntilDestroy({ checkProperties: true })
 @Component({
 	selector: 'ngx-settings',
@@ -52,7 +57,7 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 		this._logContents$.next(logs);
 	}
 	logIsOpen = false;
-	isRestart = false;
+	private _isRestart$: BehaviorSubject<boolean>;
 
 	appName: string = this.electronService.remote.app.getName();
 	menus = [];
@@ -384,6 +389,7 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 	minimizeOnStartup = null;
 	authSetting = null;
 	currentUser$: BehaviorSubject<any> = new BehaviorSubject(null);
+	currentUser$: BehaviorSubject<any> = new BehaviorSubject(null);
 	serverTypes = {
 		integrated: 'Integrated',
 		custom: 'Custom',
@@ -431,6 +437,7 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 		private _dialogService: NbDialogService,
 		private _setupService: SetupService,
 		private _notifier: ToastrNotificationService,
+		@Optional()
 		private _authStrategy: AuthStrategy
 	) {
 		this._loading$ = new BehaviorSubject(false);
@@ -459,6 +466,7 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 		this._restartDisable$ = new BehaviorSubject(false);
 		this._isHidden$ = new BehaviorSubject(true);
 		this._simpleScreenshotNotification$ = new BehaviorSubject(false);
+		this._isRestart$ = new BehaviorSubject(false);
 	}
 
 	ngOnInit(): void {
@@ -470,6 +478,8 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 				filter(() => !this._isCheckHost.status),
 				tap(() => this._restartDisable$.next(true)),
 				untilDestroyed(this)
+				tap(() => this._restartDisable$.next(true)),
+				untilDestroyed(this)
 			)
 			.subscribe();
 		this.isCheckHost$
@@ -477,6 +487,14 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 				tap(({ status }) => this._restartDisable$.next(!status)),
 				filter(() => !this._isConnectedDatabase.status),
 				tap(() => this._restartDisable$.next(true)),
+				untilDestroyed(this)
+			)
+			.subscribe();
+		this.isRestart$
+			.pipe(
+				tap((isRestart: boolean) =>
+					this._restartDisable$.next(isRestart)
+				),
 				untilDestroyed(this)
 			)
 			.subscribe();
@@ -523,21 +541,22 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 				this._simpleScreenshotNotification$.next(setting?.simpleScreenshotNotification);
 				this.selectedPeriod = setting?.timer?.updatePeriod;
 				if (this.isDesktopTimer) {
-					await this.getUserDetails();
-				}
-				this.menus = this.isServer
-					? ['Update', 'Advanced Setting', 'About']
-					: [
+					if (this.isDesktopTimer) {
+						await this.getUserDetails();
+					}
+					this.menus = this.isServer
+						? ['Update', 'Advanced Setting', 'About']
+						: [
 							...(auth && auth.allowScreenshotCapture ? ['Screen Capture'] : []),
 							'Timer',
 							'Update',
 							'Advanced Setting',
 							'About'
-					  ];
-				const lastMenu =
-					this._selectedMenu && this.menus.includes(this._selectedMenu) ? this._selectedMenu : this.menus[0];
-				this._selectedMenu$.next(lastMenu);
-			})
+						];
+					const lastMenu =
+						this._selectedMenu && this.menus.includes(this._selectedMenu) ? this._selectedMenu : this.menus[0];
+					this._selectedMenu$.next(lastMenu);
+				})
 		);
 
 		this.electronService.ipcRenderer.on('app_setting_update', (event, arg) =>
@@ -794,14 +813,11 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 	}
 
 	public async restartApp(): Promise<void> {
-		if (this.isServer && this.serverIsRunning) {
-			this._restartDisable$.next(true);
-		} else {
-			if (!this.authSetting.isLogout) {
-				await firstValueFrom(this._authStrategy.logout());
-				this.currentUser$.next(null);
-				localStorage.clear();
-			}
+		this._isRestart$.next(true);
+		if (!this.isServer && !this.authSetting.isLogout) {
+			await firstValueFrom(this._authStrategy.logout());
+			this.currentUser$.next(null);
+			localStorage.clear();
 		}
 		const thConfig = {};
 		this.thirdPartyConfig.forEach((item) => {
@@ -815,7 +831,6 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 		if (this.config.timeTrackerWindow) newConfig.awHost = `http://localhost:${this.config.awPort}`;
 		this.electronService.ipcRenderer.send('restart_app', newConfig);
 		this.electronService.ipcRenderer.send('save_additional_setting', thConfig);
-		if (this.isServer) this.isRestart = false;
 	}
 
 	portChange(val, type) {
@@ -982,7 +997,7 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 				break;
 		}
 		this.toastrService.show(message, `Success`, { status: arg.status });
-		this.isRestart = false;
+		this._isRestart$.next(false);
 	}
 
 	logBoxChange(e) {
@@ -1200,5 +1215,9 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 
 	public get simpleScreenshotNotification(): boolean {
 		return this._simpleScreenshotNotification$.getValue();
+	}
+
+	public get isRestart$(): Observable<boolean> {
+		return this._isRestart$.asObservable();
 	}
 }
